+++ app/src/main/java/net/kwmt27/codesearch/domain/usecase/GetEvents.kt
@@ -1,10 +1,6 @@
 package net.kwmt27.codesearch.domain.usecase
 
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.disposables.CompositeDisposable
-import io.reactivex.disposables.Disposable
-import io.reactivex.observers.DisposableSingleObserver
-import io.reactivex.schedulers.Schedulers
+import io.reactivex.Single
 import net.kwmt27.codesearch.domain.model.Event
 import net.kwmt27.codesearch.domain.repository.EventsRepository
 import javax.inject.Inject
@@ -12,32 +8,21 @@ import javax.inject.Inject
 /**
  * イベントリストを取得するUseCase
  */
-class GetEvents @Inject constructor(private val eventRepository: EventsRepository) : UseCase{
+class GetEvents @Inject constructor(
+        private val eventRepository: EventsRepository
+) : UseCase<GetEvents.Companion.Params, List<Event>>() {
 
 
     companion object {
         data class Params(val user:String, val page:Int)
     }
 
-    private val disposables = CompositeDisposable()
 
-    fun execute(observer: DisposableSingleObserver<List<Event>>, params: Params) {
-        val observable = this.eventRepository.events(params.user, params.page)
-                .subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-        addDisposable(observable.subscribeWith(observer))
+    override fun buildRepository(param: Params): Single<List<Event>> {
+        return eventRepository.events(param.user, param.page)
     }
 
 
 
-    private fun addDisposable(disposable: Disposable) {
-        disposables.add(disposable)
-    }
-
-    fun dispose() {
-        if(!disposables.isDisposed){
-            disposables.dispose()
-        }
-    }
 
 }
