+++ app/src/main/java/net/kwmt27/codesearch/domain/usecase/UseCase.kt
@@ -1,4 +1,34 @@
 package net.kwmt27.codesearch.domain.usecase
 
-interface UseCase {
+import io.reactivex.Single
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.disposables.CompositeDisposable
+import io.reactivex.disposables.Disposable
+import io.reactivex.observers.DisposableSingleObserver
+import io.reactivex.schedulers.Schedulers
+
+abstract class UseCase<REQUEST, RESPONSE> {
+
+    private val disposables = CompositeDisposable()
+
+    fun execute(observer: DisposableSingleObserver<RESPONSE>, params: REQUEST) {
+        val observable = buildRepository(params)
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+        addDisposable(observable.subscribeWith(observer))
+    }
+
+    abstract fun buildRepository(param: REQUEST): Single<RESPONSE>
+
+    private fun addDisposable(disposable: Disposable) {
+        disposables.add(disposable)
+    }
+
+    fun dispose() {
+        if (!disposables.isDisposed) {
+            disposables.dispose()
+        }
+    }
+
+
 }
